Refactor: Changed and improved naming of several files and folders
