+++ package.json
@@ -3,8 +3,7 @@
   "version": "0.1.3",
   "description": "Snapshot testing for React Styleguidist",
   "repository": "https://github.com/MicheleBertoli/snapguidist.git",
-  "bin": "./bin/snapguidist.js",
-  "main": "./dist/snapguidist.js",
+  "main": "./dist/index.js",
   "files": [
     "dist"
   ],
