+++ src/components/Preview.js
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react'
 import Preview from 'react-styleguidist/src/rsg-components/Preview'
-import api from '../api/api'
+import api from '../api'
 import Test from './Test'
 
 class SnapguidistPreview extends Component {
