+++ src/core/snapshot.js
