+++ src/core/updateWebpackConfig.js
